+++ compiler/src/test/java/info/teksol/mc/mindcode/compiler/ast/AstIndentedPrinterTest.java
@@ -103,6 +103,7 @@ public class AstIndentedPrinterTest extends AbstractAstBuilderTest {
                                         inModifier = false,
                                         outModifier = false,
                                         refModifier = false,
+                                        constModifier = false,
                                         varargs = false
                                     },
                                     body = AstIfExpression {